TFS running
I was able to get the server running from downloading the tfs-v1.4.2-windows-vcpkg.zip.

I was not able to get it to work from following the compiling guide, it was not able to retrieve the lua.config when I ran the x64 release from VS.

I'm not sure why the date/time is messed up.